# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Model Computer Vision API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List all registered models",
        "description": "Returns every active model with its version, task, measured metrics and known limitations, plus which version currently serves each task.",
        "operationId": "list_models_api_v1_models_get",
        "parameters": [
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one task.",
              "title": "Task"
            },
            "description": "Filter to one task."
          },
          {
            "name": "loaded_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only models currently in memory.",
              "default": false,
              "title": "Loaded Only"
            },
            "description": "Only models currently in memory."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{name}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get one model's metadata",
        "description": "Get metadata for one model version.",
        "operationId": "get_model_api_v1_models__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Registry model name.",
              "title": "Name"
            },
            "description": "Registry model name."
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Version, or 'latest'.",
              "default": "latest",
              "title": "Version"
            },
            "description": "Version, or 'latest'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDescriptor"
                }
              }
            }
          },
          "404": {
            "description": "No such model or version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/reload": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Reload the model registry from disk",
        "description": "Re-reads registry.json so a newly registered model version becomes available without restarting the service. Requires the 'admin' scope.",
        "operationId": "reload_registry_api_v1_models_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reload Registry Api V1 Models Reload Post"
                }
              }
            }
          },
          "403": {
            "description": "Admin scope required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code, e.g. 'IMAGE_TOO_LARGE'."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Friendly, user-facing explanation."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Inner object of the standard error envelope."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error shape every failing endpoint returns."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelDescriptor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "task": {
            "$ref": "#/components/schemas/TaskType"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeFormat"
          },
          "device": {
            "type": "string",
            "title": "Device"
          },
          "loaded": {
            "type": "boolean",
            "title": "Loaded",
            "description": "True when the model is resident in memory and ready."
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "True when this serves requests that do not pin a version."
          },
          "num_classes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Classes"
          },
          "input_shape": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Shape",
            "description": "Expected input tensor shape, e.g. [1, 3, 224, 224]."
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "description": "Known failure modes, copied from the model card."
          },
          "registered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "task",
          "runtime",
          "device",
          "loaded",
          "is_default"
        ],
        "title": "ModelDescriptor",
        "description": "One entry in ``GET /api/v1/models``."
      },
...
      "ModelMetrics": {
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy",
            "description": "Top-1 accuracy, 0 to 1."
          },
          "top5_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top5 Accuracy",
            "description": "Top-5 accuracy, 0 to 1."
          },
          "map50": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map50",
            "description": "Detection mAP at IoU 0.50."
          },
          "map50_95": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map50 95",
            "description": "Detection mAP averaged over IoU .50-.95."
          },
          "recall_at_1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall At 1",
            "description": "Similarity retrieval Recall@1."
          },
          "p50_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Latency Ms",
            "description": "Median measured latency."
          },
          "p95_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Latency Ms",
            "description": "95th percentile measured latency."
          },
          "throughput_rps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throughput Rps",
            "description": "Measured requests per second."
          },
          "size_mb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Mb",
            "description": "On-disk artifact size."
          }
        },
        "type": "object",
        "title": "ModelMetrics",
        "description": "Evaluation and runtime metrics for a registered model."
      },
      "ModelsResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelDescriptor"
            },
            "type": "array",
            "title": "Models"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "defaults": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Defaults",
            "description": "Default 'name:version' chosen for each task when none is pinned."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "models",
          "count",
          "defaults",
          "correlation_id"
        ],
        "title": "ModelsResponse",
        "description": "Response for ``GET /api/v1/models``."
      },
      "RuntimeFormat": {
        "type": "string",
        "enum": [
          "torch",
          "torch_int8",
          "onnx",
          "onnx_int8",
          "tensorrt"
        ],
        "title": "RuntimeFormat",
        "description": "Which compiled form of a model is executing."
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "classification",
          "detection",
          "similarity"
        ],
        "title": "TaskType",
        "description": "The three computer-vision tasks this system serves."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued to your account. Determines your tier."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
}